Attachments become a single direction link.
Attachments knew about the entities that linked to it as well as the entities knowing the set of attachments they had. The link only need to be from the entities to the attachments. We drop some explicit indexes but the foreign key constrains generate the same indexes.

This breaks some tool functionality as the session won’t be open so calls will fail.