Removed more fields from attachment.
Also removed code from the tool as removing an attachment isn’t done from another screen any longer.
I *think* the getPvtMsgAttachId was used so they attachment could have an ID before it was persisted to the DB so they could be added and removed without having to persist them. All wrong to put this into the API.